Removes default rake task additions and adds a 'ci' task to run all spec targets.
Kills simulator when retrying to run a grep_cmd_for_failure (that is when we get a 256 status code from the simulator).

Fixes bug where running something using grep_cmd_for_failure would always exit the process regardless of success or failure.

We now reenable the cedar_spec and build_specs tasks before running them so that we can use these tasks with the list of spec targets in the ci task (because using invoke usually only lets you run a task once).