replaces 'private' rake tasks with methods on ThrustConfig
Previously we were reenabling and invoking tasks that were really not meant to be run on the command line.  Now they are methods on ThrustConfig (not sure if it is it's final resting ground) that can be re-used by multiple rake tasks.
Most notably we have extracted  xcodeclean(build_configuration, sdk) and xcodebuild(build_configuration, sdk, target) as the interface the the xcodebuild CLI

Additionally we no longer try to clean some pre-defined set of build configurations every time you run specs but instead clean the specific configuration your are running in your spec. We will also clean the sdk type (iphoneos/iphonesimulator) you are attempting to build instead of always cleaning iphoneos only.