Building an ipa should not always go back to master
Rather than checking out master, we should checkout the last branch or sha that the
user was at. The use case for this is building an ipa from a feature branch, if thrust
executes "git checkout master" at the end, then the user is no longer at their
original checkout point after we checkout the deployment tag.